requirements.txt
python-telegram-bot
sastrawi